Role Registration
I have added functionality to the register page that enables users to select the role that they would like to have. This is for testing purposes and can be changed if necessary. This marks the completion of part 1. I will now move on to adding user profiles and completing part 3.